拆分main.py中测试节点的相关代码到新增的tester.py模块中
